signal-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
slowapi
pytest
httpx==0.27.2
google-generativeai
requests
firebase-admin
python-dotenv
